# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MathRAG THPT API",
    "description": "Grounded RAG API for Vietnamese high-school mathematics.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/documents": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_api_v1_admin_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDocumentList"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/documents/upload": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_admin_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_admin_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "operationId": "chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Stream Chat",
        "operationId": "stream_chat_api_v1_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/quizzes/generate": {
      "post": {
        "tags": [
          "quizzes"
        ],
        "summary": "Generate Quiz",
        "operationId": "generate_quiz_api_v1_quizzes_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/quizzes/{quiz_id}/submit": {
      "post": {
        "tags": [
          "quizzes"
        ],
        "summary": "Submit Quiz",
        "operationId": "submit_quiz_api_v1_quizzes__quiz_id__submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quiz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quiz Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizSubmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminDocumentList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminDocumentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "AdminDocumentList"
      },
      "AdminDocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "grade": {
            "type": "integer",
            "title": "Grade"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "uploaded_by": {
            "type": "string",
            "title": "Uploaded By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "original_filename",
          "grade",
          "topic",
          "content_type",
          "chunk_count",
          "uploaded_by",
          "created_at"
        ],
        "title": "AdminDocumentResponse"
      },
      "AssistanceLevel": {
        "type": "string",
        "enum": [
          "hint",
          "guided",
          "full"
        ],
        "title": "AssistanceLevel"
      },
      "AuthResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "AuthResponse"
      },
      "Body_upload_document_api_v1_admin_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 2,
            "title": "Title"
          },
          "topic": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Topic"
          },
          "grade": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 10.0,
            "title": "Grade",
            "default": 12
          },
          "content_type": {
            "type": "string",
            "pattern": "^(theory|formula|example|exam|solution)$",
            "title": "Content Type",
            "default": "theory"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "file",
          "title",
          "topic"
        ],
        "title": "Body_upload_document_api_v1_admin_documents_upload_post"
      },
      "ChatMode": {
        "type": "string",
        "enum": [
          "study",
          "exam"
        ],
        "title": "ChatMode"
      },
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 2,
            "title": "Message"
          },
          "mode": {
            "$ref": "#/components/schemas/ChatMode",
            "default": "study"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "assistance_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssistanceLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "filters": {
            "$ref": "#/components/schemas/SearchFilters"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "mode": {
            "$ref": "#/components/schemas/ChatMode"
          },
          "assistance_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssistanceLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "related_documents": {
            "items": {
              "$ref": "#/components/schemas/RelatedDocument"
            },
            "type": "array",
            "title": "Related Documents"
          },
          "detected_formulas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detected Formulas"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "fallback_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Status"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "answer",
          "mode",
          "assistance_level",
          "citations",
          "evidence",
          "related_documents",
          "detected_formulas",
          "confidence"
        ],
        "title": "ChatResponse"
      },
      "Citation": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "chunk_id",
          "title",
          "label"
        ],
        "title": "Citation"
      },
      "ContentType": {
        "type": "string",
        "enum": [
          "theory",
          "formula",
          "example",
          "exam",
          "solution"
        ],
        "title": "ContentType"
      },
...
      "Evidence": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "score_breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Score Breakdown"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "title",
          "topic",
          "content_type",
          "excerpt",
          "score"
        ],
        "title": "Evidence"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "QuizGenerateRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Topic"
          },
          "difficulty": {
            "type": "string",
            "pattern": "^(easy|medium|hard)$",
            "title": "Difficulty",
            "default": "medium"
          },
          "question_count": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Question Count",
            "default": 3
          },
          "assistance_level": {
            "$ref": "#/components/schemas/AssistanceLevel",
            "default": "hint"
          }
        },
        "type": "object",
        "required": [
          "topic"
        ],
        "title": "QuizGenerateRequest"
      },
      "QuizQuestion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "source_chunk_id": {
            "type": "string",
            "title": "Source Chunk Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "prompt",
          "options",
          "source_chunk_id"
        ],
        "title": "QuizQuestion"
      },
      "QuizResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "assistance_level": {
            "$ref": "#/components/schemas/AssistanceLevel"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestion"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "topic",
          "difficulty",
          "assistance_level",
          "questions"
        ],
        "title": "QuizResponse"
      },
      "QuizResult": {
        "properties": {
          "quiz_id": {
            "type": "string",
            "title": "Quiz Id"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/QuizResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "quiz_id",
          "score",
          "total",
          "results"
        ],
        "title": "QuizResult"
      },
      "QuizResultItem": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "correct_option": {
            "type": "integer",
            "title": "Correct Option"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "correct",
          "correct_option",
          "explanation"
        ],
        "title": "QuizResultItem"
      },
      "QuizSubmission": {
        "properties": {
          "answers": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "QuizSubmission"
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Full Name"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "password"
        ],
        "title": "RegisterRequest"
      },
      "RelatedDocument": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "topic",
          "content_type"
        ],
        "title": "RelatedDocument"
      },
      "SearchFilters": {
        "properties": {
          "grade": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade",
            "default": 12
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics"
          },
          "content_types": {
            "items": {
              "$ref": "#/components/schemas/ContentType"
            },
            "type": "array",
            "title": "Content Types"
          }
        },
        "type": "object",
        "title": "SearchFilters"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active"
        ],
        "title": "UserResponse"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "user",
          "admin"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}